Add validator methods to new file
