+++ mocks/media/withDuplicatedBreaks.m3u8
@@ -12,7 +12,7 @@ channel-audio_1=96000-video=3442944-546178966.ts
 #EXT-X-DATERANGE:ID="4026554821-1747772692",START-DATE="2025-05-20T20:24:52.699999Z",PLANNED-DURATION=20,SCTE35-OUT=0xFC3025000000000BB800FFF01405F00059C67FEFFF253C6500FE001B7740000101010000CB09D539
 #EXT-X-CUE-OUT:20
 ## splice_insert(auto_return)
-#EXT-X-DATERANGE:ID="4026554822-1747772692",START-DATE="2025-05-20T20:24:52.699999Z",PLANNED-DURATION=20,SCTE35-OUT=0xFC3025000000000BB800FFF01405F00059C67FEFFF253C6500FE001B7740000101010000CB09D539
+#EXT-X-DATERANGE:ID="4026554822-1747772692",START-DATE="2025-05-20T20:24:53.122222Z",PLANNED-DURATION=20,SCTE35-OUT=0xFC3025000000000BB800FFF01405F00059C67FEFFF253C6500FE001B7740000101010000CB09D539
 #EXT-X-CUE-OUT:20
 #EXT-X-PROGRAM-DATE-TIME:2025-05-20T20:24:52.699999Z
 #EXTINF:2.1333, no desc
